Move Category model to own app
