+++ backend/api/categories/serializers.py
@@ -1,0 +1,18 @@
+from rest_framework import serializers
+
+from categories.models import Category
+
+
+class CategorySerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Category
+        fields = ["id", "name", "description", "question_count"]
+
+
+class CategoryStringSerializer(serializers.ModelSerializer):
+    def to_representation(self, value):
+        return value.name
+
+    class Meta:
+        model = Category
+        fields = ["name"]
