+++ backend/api/management/commands/import_questions_from_notion.py
@@ -6,7 +6,8 @@ from django.db import IntegrityError
 from django.utils import timezone
 
 from api import constants, utilities_notion
-from api.models import Category, Question, Quiz, Tag
+from api.models import Question, Quiz, Tag
+from categories.models import Category
 from contributions.models import Contribution
 from core.models import Configuration
