+++ backend/api/management/commands/validate_questions_in_notion.py
@@ -3,7 +3,8 @@ from django.core.exceptions import ValidationError
 from django.core.management import BaseCommand
 
 from api import utilities_notion
-from api.models import Category, Question, Tag
+from api.models import Question, Tag
+from categories.models import Category
 
 
 class Command(BaseCommand):
