+++ backend/api/models.py
@@ -7,29 +7,6 @@ from django.utils.text import slugify
 from api import constants, utilities
 
 
-class Category(models.Model):
-    name = models.CharField(max_length=50, blank=False, help_text="Le nom de la catégorie")
-    name_long = models.CharField(max_length=150, blank=False, help_text="Le nom allongé de la catégorie")
-    description = RichTextField(blank=True, help_text="Une description de la catégorie")
-    created = models.DateField(auto_now_add=True, help_text="La date de création de la catégorie")
-
-    class Meta:
-        verbose_name = "Category"
-        verbose_name_plural = "Categories"
-        ordering = ["pk"]
-        constraints = [models.UniqueConstraint(fields=["name"], name="unique category name")]
-
-    def __str__(self):
-        return f"{self.name}"
-
-    @property
-    def question_count(self) -> int:
-        return self.questions.validated().count()
-
-    # Admin
-    question_count.fget.short_description = "Questions (validées)"
-
-
 class TagManager(models.Manager):
     def get_ids_from_name_list(self, tag_name_list: list):
         tag_ids = []
@@ -108,7 +85,7 @@ class Question(models.Model):
         help_text="Le type de question (QCM, V/F, ...)",
     )
     category = models.ForeignKey(
-        Category,
+        "categories.Category",
         blank=True,
         null=True,
         on_delete=models.SET_NULL,
