+++ backend/api/urls.py
@@ -3,10 +3,10 @@ from drf_spectacular.views import SpectacularAPIView, SpectacularRedocView, Spec
 from rest_framework import routers
 
 from api import views
+from api.categories.views import CategoryViewSet
 from api.contributions.views import ContributionViewSet
 from api.glossary.views import GlossaryViewSet
 from api.views import (
-    CategoryViewSet,
     QuestionDifficultyViewSet,
     QuestionLanguageViewSet,
     QuestionTypeViewSet,
