+++ backend/categories/factories.py
@@ -1,0 +1,11 @@
+import factory
+
+from categories.models import Category
+
+
+class CategoryFactory(factory.django.DjangoModelFactory):
+    class Meta:
+        model = Category
+        django_get_or_create = ["name"]
+
+    name = "Energie"
